requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
coverage==6.3
fastapi==0.68.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.7.0
python-dotenv==0.19.0
python-multipart==0.0.5
requests==2.26.0
six==1.16.0
starlette==0.14.2
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
unicorn==1.0.3
urllib3==1.26.8
uvicorn==0.15.0
